Potential UTC time fix and script for generating lat lon csv (find a map software)

+++ Core/Src/neom9n_thread.c
@@ -238,7 +238,7 @@ void neom9n_thread_entry(ULONG initial_input)
                         gps_packet.hours, gps_packet.minutes, gps_packet.seconds, gps_packet.milliseconds);
 
                 /* Unix Epoch */
-                printf("FC Time: %lu ms since 1/1/1970\n", (unsigned long)gps_epoch_ms);
+                printf("FC Time: %llu ms since 1/1/1970\n", (unsigned long long)gps_epoch_ms);
                 
                 /* Status */
                 printf("Status:\n");
